[Fix]ユーザ作成機能・docker修正
