+++ app/controllers/auth/auth_controller.rb
@@ -1,0 +1,2 @@
+class Auth::AuthController < ApplicationController
+end
