+++ app/models/store.rb
@@ -1,4 +1,6 @@
 class Store < ApplicationRecord
+	include IdGenerator
+
 	validates :manager_id, presence: true
 	validates :store_name, presence: true
 end
