+++ config/routes.rb
@@ -1,15 +1,7 @@
 Rails.application.routes.draw do
-	# 開発用のルーティング
-	get '/validate', to: 'auth#validate_access_token'
-	get '/fetch_events', to: 'auth#fetch_events'
-	get '/get_token', to: 'users#create'
-
-	namespace 'api' do
-		namespace 'v1' do
-			namespace 'user' do
-				get '/create', to: 'users#create'
-				get '/update', to: 'users#update'
-			end
-		end
+	namespace 'user' do
+		get '/create', to: 'users#create'
+		# 開発用
+		get '/get_user_info', to: 'auth#get_user_info'
 	end
 end
